web3_service/requirements.txt
flask==3.0.3
flask-cors>=4.0.1
sqlalchemy==2.0.9
pymysql==1.1.0
cryptography==42.0.5
hdwallet==2.2.1
solders==0.4.0
mnemonic==0.21
ccxt==4.5.32
